implement decoder bit reader layer
GenericReader is a very powerful interface which allows for arbitrary
byte manipulation behavior layering.

In this case, I was able to wrap a normal byte reader such that every
two decoded words are also byte-packed. This allows bitreader to
properly the bit stream as if it were in words (which is not normally
doable due to limitation with @memcpy and non-byte-sized types like u4).